road-capacity-api/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.7
pydantic==1.10.7
geoalchemy2==0.12.5
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==1.0.0
apscheduler==3.10.1
shapely==2.0.1
python-multipart==0.0.6
uuid==1.30